pyproject.toml
[project]
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: Software Development :: Testing",
]
description = "Patternfly5 widget library for Widgetastic."
dynamic = ["version"]
keywords = ["widgetastic", "patternfly", "patternfly5"]
license = {file = "LICENSE"}
maintainers = [
  {name = "Nikhil Dhandre", email = "ndhandre@redhat.com"},
  {name = "Egor Shamardin", email = "eshamard@redhat.com"},
  {name = "Mike Shriver", email = "mshriver@redhat.com"},
  ]
name = "widgetastic.patternfly5"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "widgetastic.core>=1.0.6",
]

[project.optional-dependencies]
dev = [
  "pre-commit",
  "pytest",
  "pytest-cov",
  "pytest-xdist",
  "pytest-rerunfailures",
  "codecov",
]
doc = ["sphinx"]

[project.urls]
repository = "https://github.com/RedHatQE/widgetastic.patternfly5"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling", "hatch-vcs"]

[tool.hatch.build.targets.wheel]
packages = ["src/widgetastic_patternfly5"]
strict-naming = false

[tool.hatch.version]
source = "vcs"
raw-options.version_scheme = "calver-by-date"

[tool.ruff]
line-length = 100
extend-ignore = ["PLR2004", "PLE0643", "PLR0913"]

[tool.ruff.lint]
select = ["F", "I", "PL", "UP"]
